delete useless code for chat-evaluator
