+++ app/toastmasters/chat-input.tsx
@@ -1,34 +1,11 @@
 import { useDebouncedCallback } from "use-debounce";
-import React, {
-  useState,
-  useRef,
-  useEffect,
-  useMemo,
-  useCallback,
-} from "react";
+import React, { useState, useRef, useEffect } from "react";
 
 import MicphoneIcon from "../icons/Micphone.svg";
 
-import {
-  ChatMessage,
-  SubmitKey,
-  useChatStore,
-  BOT_HELLO,
-  createMessage,
-  useAccessStore,
-  Theme,
-  useAppConfig,
-  DEFAULT_TOPIC,
-  ALL_MODELS,
-} from "../store";
+import { useAppConfig } from "../store";
 
-import {
-  copyToClipboard,
-  downloadAs,
-  selectOrCopy,
-  autoGrowTextArea,
-  useMobileScreen,
-} from "../utils";
+import { autoGrowTextArea } from "../utils";
 
 import { IconButton } from "../components/button";
 import styles from "../components/chat.module.scss";
@@ -141,6 +118,55 @@ export const ChatInput = (props: {
           }
           onClick={onRecord}
         />
+      </div>
+    </div>
+  );
+};
+
+export const ChatAction = (props: {
+  text: string;
+  icon: JSX.Element;
+  onClick: () => void;
+}) => {
+  const iconRef = useRef<HTMLDivElement>(null);
+  const textRef = useRef<HTMLDivElement>(null);
+  const [width, setWidth] = useState({
+    full: 16,
+    icon: 16,
+  });
+
+  function updateWidth() {
+    if (!iconRef.current || !textRef.current) return;
+    const getWidth = (dom: HTMLDivElement) => dom.getBoundingClientRect().width;
+    const textWidth = getWidth(textRef.current);
+    const iconWidth = getWidth(iconRef.current);
+    setWidth({
+      full: textWidth + iconWidth,
+      icon: iconWidth,
+    });
+  }
+
+  return (
+    <div
+      className={`${styles["chat-input-action"]} clickable`}
+      onClick={() => {
+        props.onClick();
+        setTimeout(updateWidth, 1);
+      }}
+      onMouseEnter={updateWidth}
+      onTouchStart={updateWidth}
+      style={
+        {
+          "--icon-width": `${width.icon}px`,
+          "--full-width": `${width.full}px`,
+        } as React.CSSProperties
+      }
+    >
+      <div ref={iconRef} className={styles["icon"]}>
+        {props.icon}
+      </div>
+      <div className={styles["text"]} ref={textRef}>
+        {props.text}
       </div>
     </div>
   );
